chore:  redesign trade screen
